// client-app/src/app/core/enums/activity/activity-location.enum.ts
export enum ActivityLocation {
  // Admin page
  Admin = 1,
  // Admit users
  AdmitUsers = 2,
  // Review user
  ReviewUser = 3,
  // User management
  UserManagement = 4,
  // Add user
  AddUser = 5,
  // Edit user
  EditUser = 6,
  // Company management
  CompanyManagement = 7,
  // Add company
  AddCompany = 8,
  // Edit company
  EditCompany = 9,
  // Group management
  GroupManagement = 10,
  // Add group
  AddGroup = 11,
  // Edit group
  EditGroup = 12,
  // Learn page
  Learn = 13,
  // Learn details
  LearnDetails = 14,
  // Events page
  Events = 15,
  // Event details
  EventDetails = 16,
  // Add event
  AddEvent = 17,
  // Forums page
  Forums = 18,
  // Forum details
  ForumDetails = 19,
  // Add forum
  AddForum = 20,
  // Tool management
  ToolManagement = 21,
  // Tools page
  Tools = 22,
  // Add tool
  AddTool = 23,
  // Edit tool
  EditTool = 24,
  // View tool
  ViewTool = 25,
  // Email alert settings
  EmailAlertSettings = 26,
  // Announcement management
  AnnouncementManagement = 27,
  // Add announcement
  AddAnnouncement = 28,
  // Edit announcement
  EditAnnouncement = 29,
  // Dashboard page
  Dashboard = 30,
  // Project catalog
  ProjectCatalog = 31,
  // Project details
  ProjectDetails = 32,
  // Project library
  ProjectLibrary = 33,
  // Add project
  AddProject = 34,
  // Edit project
  EditProject = 35,
  // Community page
  Community = 36,
  // Search result page
  SearchResult = 37,
  // Notifications page
  Notifications = 38,
  // Messages page
  Messages = 39,
  // Add message
  AddMessage = 40,
  // View message
  ViewMessage = 41,
  // Saved content page
  SavedContent = 42,
  // Topic page
  Topic = 43,
  // View user profile
  ViewUserProfile = 44,
  // Edit user profile
  EditUserProfile = 45,
  // View company profile
  ViewCompanyProfile = 46,
  // Account settings
  AccountSettings = 47,
  // Account settings
  RegistrationPage = 48,
  // Solution Details Page
  SolutionDetails = 49,
  //Solution Types Page
  SolutionTypes = 50,
  //Manage Page
  Manage = 50,
  // View Inititative Page,
  ViewInitiative = 52,
  //Create Initiative,
  CreateInitiative = 53,
  //Initiative Manage Module Page,
  InitiativeManageModulePage = 54
}