// client-app/src/app/core/enums/activity/activity-type.enum.ts
export enum ActivityTypeEnum {
  // Navigation menu item click
  NavMenuItemClick = 1,
  // Search filter apply
  SearchApply = 2,
  // Specific filter apply
  FilterApply = 3,
  // Company profile open
  CompanyProfileView = 4,
  // Follow a company
  CompanyFollow = 5,
  // Contact Provider button click
  ContactProviderClick = 6,
  // Nevermind button click during contacting provider
  ContactProviderNevermindButtonClick = 7,
  // Confirm button click during contacting provider
  ContactProviderConfirmButtonClick = 8,
  // Event details open
  EventDetailsView = 9,
  // Attending yes/no button click
  EventAttendingButtonClick = 10,
  // First click on an element on dashboard
  FirstDashboardClick = 11,
  // Start following a user
  UserFollow = 12,
  // Start following a forum
  ForumFollow = 13,
  // Add a forum to saved items
  ForumSave = 14,
  // Forum open
  ForumView = 15,
  // Click on a link
  LinkClick = 16,
  // Add a project to saved items
  ProjectSave = 17,
  // Project details open
  ProjectView = 18,
  // Click on a tool
  ToolClick = 19,
  // User profile open
  UserProfileView = 20,
  // Start a Discussion button click
  NewDiscussionClick = 21,
  // View Map button click
  ViewMapClick = 22,
  // Connect with NEO button click
  ConnectWithNEOClick = 23,
  // New Project button click
  NewProjectClick = 24,
  // Response on a forum message
  ForumMessageResponse = 25,
  // Forum message like button click
  ForumMessageLike = 26,
  // Article open
  LearnView = 27,
  // About Technologies/Solutions button click
  TechnologiesSolutionsClick = 28,
  // View all button click
  ViewAllClick = 29,
  // Project resource click
  ProjectResourceClick = 30,
  // Link button click
  LinkButtonClick = 31,
  //Announcement Button Click
  AnnouncementButtonClick = 32,
  //Signup Click
  SignupClick = 33,
  //Event Registration
  EventRegistration = 34,
  //Private Article Click
  PrivateLearnClick = 35,
  //Solution Details View
  SolutionDetailsView = 36,
  //Solution Type Click
  SolutionTypes = 37,
  //Create Initiative Click
  InitiativeCreate = 38,
  //Initiative Details Click
  InitiativeDetailsView = 39,
  //Initiative Sub step
  InitiativeSubstepClick = 40,
  // Initiative Module View All Click
  InitiativeModuleViewAllClick = 41,
  // Initiatives Button Vlick
  InitiativesButtonClick = 42,
  //Message Details View
  MessageDetailsView = 43,
  //Save Content To An Initiative From Attach Pop up
  SaveContentFromAttachContentPopUp
}